Rename offer app to offers
